- allow for change of output encoding with default utf-8
- avoid code duplicates by moving print and write functions to main helpers module
- remove dedicated helpers modules for xodr and xosc
- adapt imports to main helper module
