+++ scenariogeneration/xodr/__init__.py
@@ -1,9 +1,8 @@
 # __init__.py
-""" the xosc module contains the classes used to generate OpenDRIVE (xml) files. 
+""" the xodr module contains the classes used to generate OpenDRIVE (xml) files. 
 
 """
 from .geometry import *
-from .helpers import *
 from .opendrive import *
 from .lane import *
 from .enumerations import *
