+++ scenariogeneration/xosc/__init__.py
@@ -4,7 +4,6 @@
 """
 from .utils import *
 from .actions import *
-from .helpers import *
 from .position import *
 from .triggers import *
 from .scenario import *
